Fix timerTextView
The display time function didn't work well.
An alarm was ringing at the right time.
But the timer didn't display the right time. It's later than the exact time.
So I set an observerForever method for when lifeCycle is missing.
It works well in ADV but dodn't work on my phone(Galaxy s10e).